requirements.txt
fastapi
uvicorn[standard]
websockets
aiohttp
cryptography
aiosqlite
pyyaml
watchdog
pydantic
requests
